Fix: different user info error

+++ server/controller/reviews.js
@@ -25,32 +25,31 @@ module.exports = {
       userInfo['userImage'] = userInfo['image'];
       delete userInfo['image'];
 
-      if (!hasReview) {
-        if (Object.keys(req.files).length !== 0) {
-          let review = await Review.create({
-            content,
-            rating,
-            image1: req.files.img1 ? req.files.img1[0].location : null,
-            image2: req.files.img2 ? req.files.img2[0].location : null,
-            image3: req.files.img3 ? req.files.img3[0].location : null,
-            video: req.files.video ? req.files.video[0].location : null,
-            userId,
-            keyboardId,
-          });
-          Object.assign(review.dataValues, userInfo);
-          return res.status(200).json({ data: review });
-        } else {
-          let review = await Review.create({
-            content,
-            rating,
-            userId,
-            keyboardId,
-          });
-          Object.assign(review.dataValues, userInfo);
-          return res.status(200).json({ data: review });
-        }
+      if (hasReview)
+        return res.status(409).json({ message: '이미 리뷰를 남기셨습니다.' });
+
+      if (Object.keys(req.files).length !== 0) {
+        let review = await Review.create({
+          content,
+          rating,
+          image1: req.files.img1 ? req.files.img1[0].location : null,
+          image2: req.files.img2 ? req.files.img2[0].location : null,
+          image3: req.files.img3 ? req.files.img3[0].location : null,
+          video: req.files.video ? req.files.video[0].location : null,
+          userId,
+          keyboardId,
+        });
+        Object.assign(review.dataValues, userInfo);
+        return res.status(200).json({ data: review });
       } else {
-        return res.sendStatus(409);
+        let review = await Review.create({
+          content,
+          rating,
+          userId,
+          keyboardId,
+        });
+        Object.assign(review.dataValues, userInfo);
+        return res.status(200).json({ data: review });
       }
     } catch (error) {
       console.log(error);
@@ -73,6 +72,9 @@ module.exports = {
       },
       raw: true,
     });
+
+    if (!review)
+      return res.status(409).json({ message: '남긴 리뷰가 없습니다.' });
 
     const userInfo = await User.findOne({
       attributes: ['nickname', 'image'],
@@ -143,6 +145,16 @@ module.exports = {
     const keyboardId = req.params.id;
     const userId = req.userId;
     try {
+      const hasReview = await Review.findOne({
+        where: {
+          userId,
+          keyboardId,
+        },
+      });
+
+      if (!hasReview)
+        return res.status(409).json({ message: '남긴 리뷰가 없습니다.' });
+
       await Review.destroy({
         where: {
           userId,
